Stash unsaved changes as normal commits
Sometimes a user has unsaved changes in the index or the working tree.
We can think of these changes as two anonymous commits on the end of the
current chain.  All operations on the patch chain should treat these
patches the same way they treat any other patches.

For example, once we improve merge conflict resolution we want those
improvements to assist with stash conflicts too.

As an added bonus, it gets away from "git stash create" which was very
hard to recover from in case of unexpected errors.