Teach 'float' about index and worktree commits
The float command used to assume that the floated patch should go at the
very end of the commit chain.  Now it accounts for the fact that the
final two commits on the chain are the saved index and the saved
worktree.  The floated commit should come immediately before those.