Allow empty commits when cherry picking
Now that we stash the index and worktree as commits, it's very possible
that these commits will be empty.  We want them to be cherry picked even
though they're empty.

This doesn't change the way that redundant commits are removed by
cherry-pick.  That still happens and it's still what we want.  This only
affects commits that start out empty.