Teach 'save to' about index and worktree commits
The 'save to' command used to assume that the patch with fixup changes
was at the very end of the commit chain.  Now it accounts for the fact
that the final two commits on the chain are the saved index and the
saved worktree.  The commit with fixup changes comes immediately before
those.