di: disable break-rewrites
In my day to day usage, this option is less helpful than omitting it.
If I want to add it later, I should probably tweak the percentages so it
triggers at a higher threshold than the default.

The biggest example where --break-rewrites misbehaved is in a large
clean up patch in which code is only deleted.  That impacts more than
70% of the file but it's clearly not the use case for which this option
was intended.