Move code into prolog/tp.pl
I tend to prefer when command line tools are implemented as a thin
wrapper on top of a reusable library.  This is the first step in that
direction.  For now, the "library" is just the exact same code as it was
before.

The .pl extension on prolog/tp.pl also allows GitHub to recognize that
this is Prolog code.  Now it can provide syntax highlighting and
accurate language metrics.