Compile command line tools into saved states
When one ran 'tp' before this commit, it was locating, parsing and
compiling a bunch of Prolog files.  Now it loads a single saved state
file.

This gives faster load times (avoids locating, parsing and compiling).
It also creates a single binary which I might be able to distribute to
others who don't want to build from source or install SWIPL.