Add use_module declarations
I used to omit the declarations, relying on autolading to find the code.
That made sense when users ran the raw Prolog and didn't want to pay a
penalty for loading code that wasn't used.  However, now that we build a
saved state, we want to load as much code as possible during
compilation.