Use module-prefix notation for imported predicates
Instead of importing a predicate definition into the current module's
namespace, leave the predicate in the module where it was defined and
use the module name as a prefix.

This calling syntax is non-standard in the Prolog community but I find
it helpful.  It avoids a busy namespace.  It also reminds me, at the
call site, which module I'm dealing with.  Maybe I've been writing too
much Go code lately :-)