Start towards merging with patch algebra
The first step of performing a patch algebra merge is to invert each
patch on the right branch until we get back the same content as the
origin.  Then we append all the patches from the left branch.  No we
have the same content as the left branch.

The final steps (not implemented here) are: commute inverse patches to
the top, drop them.  Now we have all the changes from the left and right
branches in one sequence.