Work with patches in "git order"
There are two ways to represent a patch series as a list.  In "execution order"
the head of a patch series is applied then its tail.  In "git order" the
head of a patch series is the final commit; the tail is those before it.

Git order is named that because it's the order you see patches when
running "git log".  It's also the most natural way to store patches when
performing algebra on them because patch algebra allows us to drop
patches on the tip without further effort.