Portray patch series in canonical order
This defines a canonical order for a patch series in which operations on
small numbered lines are executed first.  Within the same line, content
removal happens first.  This produces a patch series in an order that's
similar to how "diff -u" would do it.

As an added benefit, the canonical order removes patches which cancel
each other out.  It can be viewed as a primitive patch simplification
process.

Aside from giving users a better visual rendering of the patch series, I
suspect that the canonical order will also make it faster to apply and
refactor (split into smaller, semantic patches) a patch series.

The sort algorithm is effectively a bubble sort but patch series are
usually almost sorted so the performance seems OK.