Rename *-main.pl to *-cli.pl
When someone first encountered the codebase, they were likely to think
that *-main.pl contained the main code of the project.  That's not so.
Rather those files define the command line tools.  Using the name "-cli"
suggests that more readily.

This also makes a better distinction between what I hope will eventually
be a cleanly separated CLI and a library.